❗️ Fix     : pnpm build 오류 수정
